// frontend/contracts/core/v1/models/ContestAnnouncementModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ContestAnnouncementModel = {
    id: string;
    contest_id: string;
    problem_id?: string;
    problem_title?: string;
    problem_letter?: string;
    author_id: string;
    author_username: string;
    title: string;
    body: string;
    created_at: string;
    updated_at: string;
};