// frontend/contracts/core/v1/models/ContestProblemListItemModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ContestProblemListItemModel = {
    problem_id: string;
    position: number;
    title: string;
    memory_limit: number;
    time_limit: number;
    created_at: string;
    updated_at: string;
};